refactor: convert spectator context to ports and adapters; extract test fixtures
