+++ src/main/java/com/wkaiser/riotapimcpserver/spectator/adapter/in/mcp/LiveGameTool.java
@@ -1,8 +1,8 @@
-package com.wkaiser.riotapimcpserver.riot.lol.spectator.tool;
+package com.wkaiser.riotapimcpserver.spectator.adapter.in.mcp;
 
 import com.wkaiser.riotapimcpserver.spectator.domain.CurrentGameInfo;
 import com.wkaiser.riotapimcpserver.spectator.domain.FeaturedGames;
-import com.wkaiser.riotapimcpserver.riot.lol.spectator.service.SpectatorService;
+import com.wkaiser.riotapimcpserver.spectator.application.SpectatorService;
 import com.wkaiser.riotapimcpserver.summoner.domain.Summoner;
 import com.wkaiser.riotapimcpserver.summoner.application.SummonerService;
 import com.wkaiser.riotapimcpserver.shared.enums.RiotApiPlatformUri;
