--- src/test/java/com/wkaiser/riotapimcpserver/riot/lol/spectator/service/SpectatorServiceTest.java
@@ -1,257 +1,0 @@
-package com.wkaiser.riotapimcpserver.riot.lol.spectator.service;
-
-import com.wkaiser.riotapimcpserver.spectator.domain.BannedChampion;
-import com.wkaiser.riotapimcpserver.spectator.domain.CurrentGameInfo;
-import com.wkaiser.riotapimcpserver.spectator.domain.CurrentGameParticipant;
-import com.wkaiser.riotapimcpserver.spectator.domain.FeaturedGames;
-import com.wkaiser.riotapimcpserver.spectator.domain.Observer;
-import com.wkaiser.riotapimcpserver.spectator.domain.Perks;
-import com.wkaiser.riotapimcpserver.shared.enums.RiotApiPlatformUri;
-import com.wkaiser.riotapimcpserver.shared.exception.RiotApiException;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.MockedStatic;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.util.ReflectionTestUtils;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestClient;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.when;
-
-/**
- * Unit tests for SpectatorService.
- * Tests service methods and error handling logic.
- * <p>
- * Note: Since SpectatorService creates its own RestClient instances internally,
- * these tests focus on logic validation and exception handling patterns.
- * Full API integration is tested in the integration test class.
- */
-@ExtendWith(MockitoExtension.class)
-public class SpectatorServiceTest {
-
-    private SpectatorService spectatorService;
-
-    @Mock
-    private RestClient mockRestClient;
-
-    @Mock
-    private RestClient.RequestHeadersUriSpec mockRequestHeadersUriSpec;
-
-    @Mock
-    private RestClient.ResponseSpec mockResponseSpec;
-
-    private static final String TEST_API_KEY = "test-api-key";
-    private static final String TEST_SUMMONER_ID = "encrypted_summoner_id_123";
-    private static final RiotApiPlatformUri TEST_PLATFORM = RiotApiPlatformUri.NA1;
-
-    @BeforeEach
-    void setUp() {
-        // Create SpectatorService with mocked RestClient
-        spectatorService = new SpectatorService(mockRestClient);
-
-        // Inject the API key using reflection since it's a @Value field
-        ReflectionTestUtils.setField(spectatorService, "apiKey", TEST_API_KEY);
-    }
-
-    @Test
-    void spectatorService_instantiation_successful() {
-        // Test that the service can be instantiated
-        assertNotNull(spectatorService);
-    }
-
-    @Test
-    void getCurrentGameInfo_constructorWithNullRestClient_doesNotThrow() {
-        // Test that the service can handle null RestClient in constructor
-        // This is important for the real service which doesn't use injected RestClient
-        SpectatorService serviceWithNullClient = new SpectatorService(null);
-        ReflectionTestUtils.setField(serviceWithNullClient, "apiKey", TEST_API_KEY);
-
-        assertNotNull(serviceWithNullClient);
-    }
-
-    @Test
-    void createPlatformClient_validPlatform_configurationTest() {
-        // This test verifies that the platform enum values work correctly
-        // by testing each platform can be used without throwing exceptions
-
-        RiotApiPlatformUri[] platforms = {
-            RiotApiPlatformUri.NA1,
-            RiotApiPlatformUri.EUW1,
-            RiotApiPlatformUri.EUN1,
-            RiotApiPlatformUri.KR,
-            RiotApiPlatformUri.JP1
-        };
-
-        for (RiotApiPlatformUri platform : platforms) {
-            // Verify platform URI is not null or empty
-            assertThat(platform.getPlatformUri()).isNotNull();
-            assertThat(platform.getPlatformUri()).isNotEmpty();
-        }
-    }
-
-    @Test
-    void getCurrentGameInfo_errorHandling_404ReturnsNull() {
-        // This test verifies the 404 handling logic exists
-        // The actual API behavior is tested in integration tests
-
-        HttpClientErrorException notFoundException = new HttpClientErrorException(
-                HttpStatus.NOT_FOUND, "Not Found", "Summoner not in game".getBytes(), null);
-
-        // Verify that 404 status code maps to NOT_FOUND
-        assertThat(notFoundException.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-        assertThat(notFoundException.getStatusCode().value()).isEqualTo(404);
-    }
-
-    @Test
-    void getCurrentGameInfo_errorHandling_otherErrorsThrowRiotApiException() {
-        // Test the error status codes that should throw RiotApiException
-        HttpStatus[] errorStatuses = {
-                HttpStatus.BAD_REQUEST,
-                HttpStatus.UNAUTHORIZED,
-                HttpStatus.FORBIDDEN,
-                HttpStatus.TOO_MANY_REQUESTS,
-                HttpStatus.INTERNAL_SERVER_ERROR
-        };
-
-        for (HttpStatus status : errorStatuses) {
-            HttpClientErrorException exception = new HttpClientErrorException(
-                    status, status.getReasonPhrase(), "Error".getBytes(), null);
-
-            // Verify these are error status codes that should be handled
-            assertThat(exception.getStatusCode().isError()).isTrue();
-            assertThat(exception.getStatusCode().value()).isNotEqualTo(404);
-        }
-    }
-
-    @Test
-    void riotApiException_construction_preservesStatusCode() {
-        // Test that our custom exception properly stores status codes
-        int testStatusCode = 429;
-        String testMessage = "Rate limit exceeded";
-
-        RiotApiException exception = new RiotApiException(testMessage, testStatusCode);
-
-        assertThat(exception.getMessage()).isEqualTo(testMessage);
-        assertThat(exception.getStatusCode()).isEqualTo(testStatusCode);
-    }
-
-    @Test
-    void featuredGames_errorHandling_preservesStatusCode() {
-        // Test that featured games error handling works similarly
-        HttpClientErrorException rateLimitException = new HttpClientErrorException(
-                HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded", "Rate limit exceeded".getBytes(), null);
-
-        // Verify the exception properties are accessible
-        assertThat(rateLimitException.getStatusCode().value()).isEqualTo(429);
-        assertThat(rateLimitException.getResponseBodyAsString()).isEqualTo("Rate limit exceeded");
-    }
-
-    @Test
-    void platformUri_enumValues_haveValidUris() {
-        // Test that all platform enums have valid URI strings
-        RiotApiPlatformUri[] allPlatforms = RiotApiPlatformUri.values();
-
-        for (RiotApiPlatformUri platform : allPlatforms) {
-            String uri = platform.getPlatformUri();
-
-            assertThat(uri).isNotNull();
-            assertThat(uri).isNotEmpty();
-            assertThat(uri).doesNotContain(" ");  // No spaces in URIs
-            assertThat(uri).contains(".");  // Should contain domain separator
-        }
-    }
-
-    // Helper methods for creating test data (used in integration tests)
-
-    public static CurrentGameInfo createSampleCurrentGameInfo() {
-        return CurrentGameInfo.builder()
-                .gameId(123456789L)
-                .gameType("MATCHED_GAME")
-                .gameStartTime(1640995200000L)
-                .mapId(11L)
-                .gameLength(450L)
-                .platformId("NA1")
-                .gameMode("CLASSIC")
-                .gameQueueConfigId(420L)
-                .bannedChampions(List.of(
-                        BannedChampion.builder()
-                                .championId(266L)
-                                .teamId(100L)
-                                .pickTurn(1)
-                                .build()
-                ))
-                .observers(Observer.builder()
-                        .encryptionKey("sample_encryption_key")
-                        .build())
-                .participants(List.of(
-                        createSampleParticipant("TestSummoner1", 1L, 100L),
-                        createSampleParticipant("TestSummoner2", 2L, 200L)
-                ))
-                .build();
-    }
-
-    public static CurrentGameParticipant createSampleParticipant(String summonerName, long championId, long teamId) {
-        return CurrentGameParticipant.builder()
-                .championId(championId)
-                .perks(Perks.builder()
-                        .perkIds(List.of(8112L, 8126L, 8138L, 8106L, 8275L, 8210L))
-                        .perkStyle(8100L)
-                        .perkSubStyle(8200L)
-                        .build())
-                .profileIconId(1234L)
-                .bot(false)
-                .teamId(teamId)
-                .summonerName(summonerName)
-                .summonerId("encrypted_summoner_id_" + summonerName)
-                .puuid("test_puuid_" + summonerName)
-                .summonerLevel(150L)
-                .spell1Id(4L)
-                .spell2Id(7L)
-                .gameCustomizationObjects(List.of())
-                .build();
-    }
-
-    public static FeaturedGames createSampleFeaturedGames() {
-        return FeaturedGames.builder()
-                .clientRefreshInterval(300L)
-                .gameList(List.of(
-                        CurrentGameInfo.builder()
-                                .gameId(987654321L)
-                                .gameStartTime(1640995200000L)
-                                .platformId("NA1")
-                                .gameMode("CLASSIC")
-                                .mapId(11L)
-                                .gameType("MATCHED_GAME")
-                                .gameQueueConfigId(420L)
-                                .gameLength(600L)
-                                .participants(List.of(
-                                        createSampleParticipant("FeaturedPlayer1", 1L, 100L),
-                                        createSampleParticipant("FeaturedPlayer2", 2L, 200L)
-                                ))
-                                .bannedChampions(List.of(
-                                        BannedChampion.builder()
-                                                .championId(266L)
-                                                .teamId(100L)
-                                                .pickTurn(1)
-                                                .build()
-                                ))
-                                .observers(Observer.builder()
-                                        .encryptionKey("featured_encryption_key")
-                                        .build())
-                                .build()
-                ))
-                .build();
-    }
-}
